Move components to own folder and some small fixes
