+++ client/packages/uml-components/.eslintrc.js
@@ -1,0 +1,5 @@
+/** @type {import('eslint').Linter.Config} */
+module.exports = {
+    extends: '../../.eslintrc.js',
+    rules: {}
+};
