+++ client/packages/uml-components/src/base/index.ts
@@ -7,6 +7,5 @@
  * SPDX-License-Identifier: MIT
  *********************************************************************************/
 
-import { usePropertyPaletteWebviewDesignSystem } from '@borkdominik-biguml/uml-components';
-
-usePropertyPaletteWebviewDesignSystem();
+export * from './component';
+export * from './style';
