+++ client/packages/uml-components/src/base/style/codicon.style.ts
@@ -25,20 +25,3 @@ function convert(style: CSSStyleSheet): CSSResult {
 }
 
 export const codiconStyle = convert(codiconStyleSheet);
-export const defaultStyle = css`
-    .title {
-        text-transform: uppercase;
-    }
-
-    h4.secondary-title {
-        margin: -1em 0 1em 0;
-    }
-
-    .no-data-message {
-        opacity: 0.5;
-    }
-
-    .hint-text {
-        opacity: 0.5;
-    }
-`;
