+++ client/packages/uml-components/src/index.ts
@@ -7,4 +7,8 @@
  * SPDX-License-Identifier: MIT
  *********************************************************************************/
 
-export const DATA_CONTEXT_MENU = 'data-context-menu';
+export * from './base';
+export * from './menu';
+export * from './property-palette';
+export * from './toolkit';
+export * from './vscode/vscode-api';
